Update Modularity of Scripts
Created a python file to hold the functions required for both scripts since they are exactly the same.  This file is called In both of the scripts to import the required functions.  A function for graphing of the histograms was also added to the functions file and was implemented within the main code.

The files were renamed with underscores to allow for import of the functions.